refactor auth
- split auth controller and service into separate files
