+++ api/src/controllers/index.ts
@@ -1,4 +1,4 @@
-// import * as authController from './authController';
+import * as authController from './authController';
 // import * as userController from './userController';
 // import * as shopController from './shopController';
 // import * as productController from './productController';
@@ -6,7 +6,7 @@
 import * as auctionController from './auctionController';
 
 const controllers = {
-  //   auth: authController,
+  auth: authController,
   //   user: userController,
   //   shop: shopController,
   //   product: productController,
