+++ api/src/routes/index.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import authRoutes from './authRoutes';
+import authRoutes from './authRoute';
 import userRoutes from './userRoutes';
 import shopRoutes from './shopRoutes';
 import productRoutes from './productRoutes';
