+++ api/src/services/index.ts
@@ -1,4 +1,4 @@
-import * as authServices from './authServices';
+import * as authServices from './authService';
 import * as userServices from './userServices';
 import * as shopServices from './shopServices';
 import * as productServices from './productServices';
